chore: Move OfferHandler from utils to controller folder
